server/pyproject.toml
[project]
name = "multi-agent-knowledge-management-system"
version = "0.1.0"
description = "Multi-agent Intelligent Knowledge Management System"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "httpx>=0.28.1",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.1",
    "langchain-pinecone>=0.2.13",
    "langchain-pymupdf4llm>=0.5.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "pinecone-client>=6.0.0",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.4.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
